refactor routes for testing, and for pleasing my eyes
