+++ client/package.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-scripts": "5.0.1",
+    "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -34,15 +34,15 @@
     ]
   },
   "devDependencies": {
-    "@types/jest": "^29.2.0",
-    "cypress": "^10.10.0",
-    "sass": "^1.55.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^29.2.0",
     "@types/node": "^16.11.65",
     "@types/react": "^18.0.21",
     "@types/react-dom": "^18.0.6",
+    "cypress": "^10.10.0",
+    "sass": "^1.55.0",
     "typescript": "^4.8.4"
   }
 }
