+++ server/src/controllers/getUnknownRoute.ts
@@ -1,0 +1,5 @@
+import { Request, Response } from "express"
+
+export const getUnknownRoute = (_: Request, respose: Response) => {
+  respose.send("These are not the droids you're looking for...")
+}
